Code for the second part of the project!
A set of if statements for the string "statistic". Taking the mean and  standard deviation of temperature and growth rate columns. Finding the number of rows by taking the length of the array. Finally, finding the mean growth rate when the temperature is < 20  by creating a new array where only rows where temp is < 20 exist. Then do the same for when temp is >50.